Convert to members report
