+++ system/modules/isotope_reports/languages/en/default.xlf
@@ -26,12 +26,6 @@
       <trans-unit id="ISO_REPORT.sales_product.1">
         <source>Sales per product</source>
       </trans-unit>
-      <trans-unit id="ISO_REPORT.sales_member.0">
-        <source>Per member</source>
-      </trans-unit>
-      <trans-unit id="ISO_REPORT.sales_member.1">
-        <source>Sales per member</source>
-      </trans-unit>
       <trans-unit id="ISO_REPORT.members_total.0">
         <source>Total</source>
       </trans-unit>
